Move and unify parameter header file into include path
